listo

+++ app/Http/Controllers/admin/UsuariosController.php
@@ -26,8 +26,18 @@ class UsuariosController extends Controller
     }
     public function borrar($id)
     {
-        $evento = DB::table('plane_user')->where('id', $id)->delete();
-        return response()->json($evento);
+        $evento = DB::table('plane_user')->where('id', $id)->first();
+        if($evento->detalle==null)
+        {
+            $evento = DB::table('plane_user')->where('id', $id)->delete();
+            return response()->json('si');
+        }
+        else
+        {
+            return response()->json('no');
+        }
+
+        
     }
     public function agregar(Request $request)
     {
